requirements.txt
Flask
flask_sqlalchemy
flask_cors
python-dotenv
gunicorn
waitress
psycopg2-binary
PyJWT
